Generalize treatment of domain hints
This change broadens the focus of the draft to cover all
domain hints provided by a local network, not just PvD.
This requires compatibility with non-apex names, which
requires clients to walk up the label tree in some cases.
It also makes a large number of other changes to the
presentation.  For example, it avoids normative requirements
on what clients do with the results of validation, and focuses
only on how validation is performed.